+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Rebalance BalrondHumanoidRandomizer spawner changes.
 - Update dependencies.
 
 ## [0.6.6] - 2024-06-05